pyproject.toml
[project]
name = "langchain-mcp-multiserver-demo"
version = "0.1.0"
description = "Multi-Server MCP Integration with LangChain - Educational Demo for AIE Cohort 8"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "ipykernel>=7.0.1",
    "langchain==0.3.19",
    "langchain-mcp-adapters>=0.1.11",
    "langchain-openai==0.3.7",
    "langgraph==0.6.7",
    "mcp[cli]>=1.6.0",
    "numpy>=2.2.4",
    "openai>=1.72.0",
    "python-dotenv>=1.1.0",
    "tavily-python>=0.5.4",
]
